// models/comp_aave_yield_optimiser/src/services/compound/types.ts
export type CompoundProtocolConstants = {
  totalBorrow: bigint;
  availableLiquidity: bigint;
  totalSupply: bigint;
  decimals: bigint;
  supplyKink: bigint;
  supplyPerSecondInterestRateBase: bigint;
  supplyPerSecondInterestRateSlopeLow: bigint;
  supplyPerSecondInterestRateSlopeHigh: bigint;
};